Create depest-tree-left.cpp
You are given a binary tree having ‘N’ number of nodes. Your task is to find the deepest leaf node in the given input tree.
Note:
The deepest leaf node is the leaf node which will be the left child of some node and will be at the maximum level in the tree.
If there are multiple deepest left leaf nodes, return the node with maximum value.